chore(engine): day 3 — move DLP module to @node9/policy-engine
The DLP module is the most self-contained piece of the matcher and the
first real code move. The pure parts go into the engine package; the
single I/O wrapper (scanFilePath, which uses fs.realpathSync.native to
defend against symlink-escape attacks) stays in the proxy.

Engine — packages/policy-engine/src/dlp/index.ts:
- DLP_PATTERNS — full pattern array (AWS, GitHub, Slack, OpenAI, Stripe,
  GCP, Azure, Databricks, DigitalOcean, Doppler, HashiCorp Vault,
  HuggingFace, Postman, Pulumi, SendGrid, PEM private keys, npm tokens,
  JWT, Stripe restricted, Slack app, GitLab, npm access, Shopify,
  Linear, PlanetScale, Sentry, Grafana, etc — entire list moved verbatim).
- scanArgs(args, depth, fieldPath) — recursive arg scanner.
- scanText(text) — single-string scanner.
- redactText(text) — replace matches with [node9-redacted:<Name>].
- matchSensitivePath(resolvedPath, originalPath) — pure path matcher
  against SENSITIVE_PATH_PATTERNS. The host resolves symlinks, then
  passes the resolved path here.
- sensitivePathMatch(originalPath) — fail-closed match builder for
  the I/O error path; exposed so the proxy's wrapper produces the
  same shape.
- SENSITIVE_PATH_REGEXES — exposed read-only for hosts that need the
  raw patterns.

Proxy — src/dlp.ts:
- Now ~60 lines (was 757). Imports matchSensitivePath +
  sensitivePathMatch from the engine. scanFilePath wraps them with
  fs.realpathSync.native for symlink resolution.
- Re-exports DLP_PATTERNS, scanArgs, scanText, redactText, DlpMatch
  from the engine so every existing import path keeps working.

Why split: the engine rule is "no I/O at evaluation time." Symlink
resolution is filesystem access, so it can't live in the engine. But
the path-pattern matching IS pure, and that's the part both proxy and
SaaS need. Splitting keeps the engine pure while still letting both
sides use the same path patterns.

Verification:
- All 1378 proxy tests pass (22 skipped, no regressions).
- Engine package: typecheck clean, build succeeds.
- Proxy: typecheck clean, lint 0 errors (8 pre-existing warnings),
  format clean.

Plan reference: doc/roadmap/v1.16.1-policy-engine.md (day 3)
