+++ packages/policy-engine/src/index.ts
@@ -5,9 +5,21 @@
 // definitions live in exactly one place. No fs/process/os imports at
 // runtime — host code passes paths and state in.
 //
-// Day 2: types only. Implementations stay in node9-proxy until Day 3+.
+// Day 3: DLP module moved here. Path-resolving I/O lives in the proxy's
+// scanFilePath wrapper since that's filesystem access.
 
 /** Engine version stamped on audit entries for future drift detection. */
 export const ENGINE_VERSION = '0.1.0-alpha.0';
 
 export type { SmartCondition, SmartRule, DlpMatch, RiskMetadata } from './types';
+
+// DLP — pure scanners. Path-resolving I/O wraps matchSensitivePath in the proxy.
+export {
+  DLP_PATTERNS,
+  scanArgs,
+  scanText,
+  redactText,
+  matchSensitivePath,
+  sensitivePathMatch,
+  SENSITIVE_PATH_REGEXES,
+} from './dlp';
